style(cli): Separate directory structure
