+++ cli/src/lib.rs
@@ -1,29 +1,10 @@
-use clap::Parser;
+mod by_clap;
+
 use file_loader::{CommandLineArgsParser, FileInfo};
 use std::error::Error;
 
 pub trait CLI {
     fn eval(&self) -> Result<CLIArgs, Box<dyn Error>>;
-}
-
-#[derive(Parser)]
-pub struct AppArg {
-    /// File's Path
-    #[arg(help = "The file you want to edit")]
-    pub path: String,
-
-    //#[arg(long, default_value_t = String::from("./init.lua"))]
-    #[arg(short, long, default_value_t = String::from("./init.lua"), help = "your config file which is as config.lua")]
-    pub config: String,
-}
-
-impl CLI for AppArg {
-    fn eval(&self) -> Result<CLIArgs, Box<dyn Error>> {
-        Ok(CLIArgs {
-            text_file: self.path.clone(),
-            config: self.config.clone(),
-        })
-    }
 }
 
 pub struct CLIArgs {
